torch dep removed
