+++ docs/source/quick_start.rst
@@ -57,8 +57,6 @@ The simulation creates one ``Bridge`` per participating rank and sends chunks.
 
     # 4 Bridges in total
     world_size = 4
-    master_address = "127.0.0.1"  # localhost only works on a single machine
-    master_port = 29500
 
     # descriptio of arrays being shared
     arrays_md = {
@@ -73,14 +71,10 @@ The simulation creates one ``Bridge`` per participating rank and sends chunks.
         }
     }
 
-    from deisa.ray.comm import init_gloo_comm
+    from mpi4py import MPI
 
-    comm = init_gloo_comm(
-        world_size,
-        rank,
-        master_address,
-        master_port,
-    )
+    comm = MPI.COMM_WORLD
+    assert comm.Get_size() == world_size
 
     # this call should be repeated 4 times with a different rank
     bridge = Bridge(
